pyproject.toml
[tool.poetry]
name = "gborelpy"
version = "0.4.0"
description = "Python utilities for Gymshark python projects"
authors = ["Gregoire Borel <gregoire.borel@neoxia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
apache-beam = {extras = ["gcp"], version = "^2.57.0"}
google-api-core = "^2.19.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
